GHC.Exts.Heap.Closures

Closures

type Closure = GenClosure Box

data GenClosure b

This is the representation of a Haskell value on the heap. It reflects https://gitlab.haskell.org/ghc/ghc/blob/master/rts/include/rts/storage/Closures.h

The data type is parametrized by b: the type to store references in. Usually this is a Box with the type synonym Closure.

All Heap objects have the same basic layout. A header containing a pointer to the info table and a payload with various fields. The info field below always refers to the info table pointed to by the header. The remaining fields are the payload.

See https://gitlab.haskell.org/ghc/ghc/wikis/commentary/rts/storage/heap-objects for more information.

Constructors

ConstrClosure
A data constructor

Fields

info :: !StgInfoTable

ptrArgs :: ![b]

Pointer arguments

dataArgs :: ![Word]

Non-pointer arguments

pkg :: !String

Package name

modl :: !String

Module name

name :: !String

Constructor name

FunClosure
A function

Fields

info :: !StgInfoTable

ptrArgs :: ![b]

Pointer arguments

dataArgs :: ![Word]

Non-pointer arguments

ThunkClosure
A thunk, an expression not obviously in head normal form

Fields

info :: !StgInfoTable

ptrArgs :: ![b]

Pointer arguments

dataArgs :: ![Word]

Non-pointer arguments

SelectorClosure
A thunk which performs a simple selection operation

Fields

info :: !StgInfoTable

selectee :: !b

Pointer to the object being selected from

PAPClosure
An unsaturated function application

Fields

info :: !StgInfoTable

arity :: !HalfWord

Arity of the partial application

n_args :: !HalfWord

Size of the payload in words

fun :: !b

Pointer to a FunClosure

payload :: ![b]

Sequence of already applied arguments

APClosure
A function application

Fields

info :: !StgInfoTable

arity :: !HalfWord

Arity of the partial application

n_args :: !HalfWord

Size of the payload in words

fun :: !b

Pointer to a FunClosure

payload :: ![b]

Sequence of already applied arguments

APStackClosure
A suspended thunk evaluation

Fields

info :: !StgInfoTable

fun :: !b

Pointer to a FunClosure

payload :: ![b]

Sequence of already applied arguments

IndClosure
A pointer to another closure, introduced when a thunk is updated to point at its value

Fields

info :: !StgInfoTable

indirectee :: !b

Target closure

BCOClosure
A byte-code object (BCO) which can be interpreted by GHC's byte-code interpreter (e.g. as used by GHCi)

Fields

info :: !StgInfoTable

instrs :: !b

A pointer to an ArrWords of instructions

literals :: !b

A pointer to an ArrWords of literals

bcoptrs :: !b

A pointer to an ArrWords of byte code objects

arity :: !HalfWord

Arity of the partial application

size :: !HalfWord

The size of this BCO in words

bitmap :: ![Word]

An StgLargeBitmap describing the pointerhood of its args/free vars

BlackholeClosure
A thunk under evaluation by another thread

Fields

info :: !StgInfoTable

indirectee :: !b

Target closure

ArrWordsClosure
A ByteArray#

Fields

info :: !StgInfoTable

bytes :: !Word

Size of array in bytes

arrWords :: ![Word]

Array payload

MutArrClosure
A MutableByteArray#

Fields

info :: !StgInfoTable

mccPtrs :: !Word

Number of pointers

mccSize :: !Word

?? Closures.h vs ClosureMacros.h

mccPayload :: ![b]

Array payload Card table ignored

SmallMutArrClosure
A SmallMutableArray#

Since: ghc-heap-8.10.1

Fields

info :: !StgInfoTable

mccPtrs :: !Word

Number of pointers

mccPayload :: ![b]

Array payload Card table ignored

MVarClosure
An MVar#, with a queue of thread state objects blocking on them

Fields

info :: !StgInfoTable

queueHead :: !b

Pointer to head of queue

queueTail :: !b

Pointer to tail of queue

value :: !b

Pointer to closure

IOPortClosure
An IOPort#, with a queue of thread state objects blocking on them

Fields

info :: !StgInfoTable

queueHead :: !b

Pointer to head of queue

queueTail :: !b

Pointer to tail of queue

value :: !b

Pointer to closure

MutVarClosure
A MutVar#

Fields

info :: !StgInfoTable

var :: !b

Pointer to contents

BlockingQueueClosure
An STM blocking queue.

Fields

info :: !StgInfoTable

link :: !b

?? Here so it looks like an IND

blackHole :: !b

The blackhole closure

owner :: !b

The owning thread state object

queue :: !b

??

WeakClosure

Fields

info :: !StgInfoTable

cfinalizers :: !b

key :: !b

value :: !b

Pointer to closure

finalizer :: !b

weakLink :: !(Maybe b)

next weak pointer for the capability

TSOClosure
Representation of StgTSO: A Thread State Object. The values for what_next, why_blocked and flags are defined in Constants.h.

Fields

info :: !StgInfoTable

link :: !b

?? Here so it looks like an IND

global_link :: !b

tsoStack :: !b

stackobj from StgTSO

trec :: !b

blocked_exceptions :: !b

bq :: !b

thread_label :: !(Maybe b)

what_next :: !WhatNext

why_blocked :: !WhyBlocked

flags :: ![TsoFlags]

threadId :: !Word64

saved_errno :: !Word32

tso_dirty :: !Word32

non-zero => dirty

alloc_limit :: !Int64

tot_stack_size :: !Word32

prof :: !(Maybe StgTSOProfInfo)

StackClosure
Representation of StgStack: The 'tsoStack ' of a TSOClosure.

Fields

info :: !StgInfoTable

stack_size :: !Word32

stack size in *words*

stack_dirty :: !Word8

non-zero => dirty

stack_marking :: !Word8

IntClosure
Primitive Int

Fields

ptipe :: PrimType

intVal :: !Int

WordClosure
Primitive Word

Fields

ptipe :: PrimType

wordVal :: !Word

Int64Closure
Primitive Int64

Fields

ptipe :: PrimType

int64Val :: !Int64

Word64Closure
Primitive Word64

Fields

ptipe :: PrimType

word64Val :: !Word64

AddrClosure
Primitive Addr

Fields

ptipe :: PrimType

addrVal :: !(Ptr ())

FloatClosure
Primitive Float

Fields

ptipe :: PrimType

floatVal :: !Float

DoubleClosure
Primitive Double

Fields

ptipe :: PrimType

doubleVal :: !Double

OtherClosure
Another kind of closure

Fields

info :: !StgInfoTable

hvalues :: ![b]

rawWords :: ![Word]

UnsupportedClosure

Fields

info :: !StgInfoTable

UnknownTypeWordSizedPrimitive
A primitive word from a bitmap encoded stack frame payload

The type itself cannot be restored (i.e. it might represent a Word8# or an Int#).

Fields

wordVal :: !Word

data PrimType

Constructors

PInt

PWord

PInt64

PWord64

PAddr

PFloat

PDouble

data WhatNext

Constructors

ThreadRunGHC

ThreadInterpret

ThreadKilled

ThreadComplete

WhatNextUnknownValue Word16
Please report this as a bug

data WhyBlocked

Constructors

NotBlocked

BlockedOnMVar

BlockedOnMVarRead

BlockedOnBlackHole

BlockedOnRead

BlockedOnWrite

BlockedOnDelay

BlockedOnSTM

BlockedOnDoProc

BlockedOnCCall

BlockedOnCCall_Interruptible

BlockedOnMsgThrowTo

ThreadMigrating

WhyBlockedUnknownValue Word16
Please report this as a bug

data TsoFlags

Constructors

TsoLocked

TsoBlockx

TsoInterruptible

TsoStoppedOnBreakpoint

TsoMarked

TsoSqueezed

TsoAllocLimit

TsoFlagsUnknownValue Word32
Please report this as a bug

allClosures :: GenClosure b -> [b]

For generic code, this function returns all referenced closures.

closureSize :: Box -> Int

Get the size of the top-level closure in words. Includes header and payload. Does not follow pointers.

Since: ghc-heap-8.10.1

Stack

type StgStackClosure = GenStgStackClosure Box

data GenStgStackClosure b

A decoded StgStack with StackFrames

Stack related data structures (GenStgStackClosure, GenStackField, GenStackFrame) are defined separately from GenClosure as their related functions are very different. Though, both are closures in the sense of RTS structures, their decoding logic differs: While it's safe to keep a reference to a heap closure, the garbage collector does not update references to stack located closures.

Additionally, stack frames don't appear outside of the stack. Thus, keeping GenStackFrame and GenClosure separated, makes these types more precise (in the sense what values to expect.)

Constructors

GenStgStackClosure

Fields

ssc_info :: !StgInfoTable

ssc_stack_size :: !Word32

stack size in *words*

ssc_stack :: ![GenStackFrame b]

type StackFrame = GenStackFrame Box

data GenStackFrame b

A single stack frame

Constructors

UpdateFrame

Fields

info_tbl :: !StgInfoTable

updatee :: !b

CatchFrame

Fields

info_tbl :: !StgInfoTable

handler :: !b

CatchStmFrame

Fields

info_tbl :: !StgInfoTable

catchFrameCode :: !b

handler :: !b

CatchRetryFrame

Fields

info_tbl :: !StgInfoTable

running_alt_code :: !Word

first_code :: !b

alt_code :: !b

AtomicallyFrame

Fields

info_tbl :: !StgInfoTable

atomicallyFrameCode :: !b

result :: !b

UnderflowFrame

Fields

info_tbl :: !StgInfoTable

nextChunk :: !(GenStgStackClosure b)

StopFrame

Fields

info_tbl :: !StgInfoTable

RetSmall

Fields

info_tbl :: !StgInfoTable

stack_payload :: ![GenStackField b]

RetBig

Fields

info_tbl :: !StgInfoTable

stack_payload :: ![GenStackField b]

RetFun

Fields

info_tbl :: !StgInfoTable

retFunSize :: !Word

retFunFun :: !b

retFunPayload :: ![GenStackField b]

RetBCO

Fields

info_tbl :: !StgInfoTable

bco :: !b

always a BCOClosure

bcoArgs :: ![GenStackField b]

type StackField = GenStackField Box

data GenStackField b

Bitmap-encoded payload on the stack

Constructors

StackWord !Word
A non-pointer field

StackBox !b
A pointer field

Boxes

data Box

An arbitrary Haskell value in a safe Box. The point is that even unevaluated thunks can safely be moved around inside the Box, and when required, e.g. in getBoxedClosureData, the function knows how far it has to evaluate the argument.

Constructors

Box (Any :: Type)

areBoxesEqual :: Box -> Box -> IO Bool

Boxes can be compared, but this is not pure, as different heap objects can, after garbage collection, become the same object.

asBox :: a -> Box

This takes an arbitrary value and puts it into a box. Note that calls like

asBox (head list)

will put the thunk "head list" into the box, not the element at the head of the list. For that, use careful case expressions:

case list of x:_ -> asBox x