GHC.Exts.Heap

Description

With this module, you can investigate the heap representation of Haskell values, i.e. to investigate sharing and lazy evaluation.

Closure types

type Closure = GenClosure Box

data GenClosure b

This is the representation of a Haskell value on the heap. It reflects http://hackage.haskell.org/trac/ghc/browser/includes/rts/storage/Closures.h

The data type is parametrized by the type to store references in. Usually this is a Box with the type synonym Closure.

All Heap objects have the same basic layout. A header containing a pointer to the info table and a payload with various fields. The info field below always refers to the info table pointed to by the header. The remaining fields are the payload.

See https://ghc.haskell.org/trac/ghc/wiki/Commentary/Rts/Storage/HeapObjects for more information.

Constructors

ConstrClosure
A data constructor

Fields

info :: !StgInfoTable

ptrArgs :: ![b]

Pointer arguments

dataArgs :: ![Word]

Non-pointer arguments

pkg :: !String

Package name

modl :: !String

Module name

name :: !String

Constructor name

FunClosure
A function

Fields

info :: !StgInfoTable

ptrArgs :: ![b]

Pointer arguments

dataArgs :: ![Word]

Non-pointer arguments

ThunkClosure
A thunk, an expression not obviously in head normal form

Fields

info :: !StgInfoTable

ptrArgs :: ![b]

Pointer arguments

dataArgs :: ![Word]

Non-pointer arguments

SelectorClosure
A thunk which performs a simple selection operation

Fields

info :: !StgInfoTable

selectee :: !b

Pointer to the object being selected from

PAPClosure
An unsaturated function application

Fields

info :: !StgInfoTable

arity :: !HalfWord

Arity of the partial application

n_args :: !HalfWord

Size of the payload in words

fun :: !b

Pointer to a FunClosure

payload :: ![b]

Sequence of already applied arguments

APClosure
A function application

Fields

info :: !StgInfoTable

arity :: !HalfWord

Arity of the partial application

n_args :: !HalfWord

Size of the payload in words

fun :: !b

Pointer to a FunClosure

payload :: ![b]

Sequence of already applied arguments

APStackClosure
A suspended thunk evaluation

Fields

info :: !StgInfoTable

fun :: !b

Pointer to a FunClosure

payload :: ![b]

Sequence of already applied arguments

IndClosure
A pointer to another closure, introduced when a thunk is updated to point at its value

Fields

info :: !StgInfoTable

indirectee :: !b

Target closure

BCOClosure
A byte-code object (BCO) which can be interpreted by GHC's byte-code interpreter (e.g. as used by GHCi)

Fields

info :: !StgInfoTable

instrs :: !b

A pointer to an ArrWords of instructions

literals :: !b

A pointer to an ArrWords of literals

bcoptrs :: !b

A pointer to an ArrWords of byte code objects

arity :: !HalfWord

Arity of the partial application

size :: !HalfWord

The size of this BCO in words

bitmap :: ![Word]

An StgLargeBitmap describing the pointerhood of its args/free vars

BlackholeClosure
A thunk under evaluation by another thread

Fields

info :: !StgInfoTable

indirectee :: !b

Target closure

ArrWordsClosure
A ByteArray#

Fields

info :: !StgInfoTable

bytes :: !Word

Size of array in bytes

arrWords :: ![Word]

Array payload

MutArrClosure
A MutableByteArray#

Fields

info :: !StgInfoTable

mccPtrs :: !Word

Number of pointers

mccSize :: !Word

?? Closures.h vs ClosureMacros.h

mccPayload :: ![b]

Array payload Card table ignored

MVarClosure
An MVar#, with a queue of thread state objects blocking on them

Fields

info :: !StgInfoTable

queueHead :: !b

Pointer to head of queue

queueTail :: !b

Pointer to tail of queue

value :: !b

Pointer to closure

MutVarClosure
A MutVar#

Fields

info :: !StgInfoTable

var :: !b

Pointer to closure

BlockingQueueClosure
An STM blocking queue.

Fields

info :: !StgInfoTable

link :: !b

?? Here so it looks like an IND

blackHole :: !b

The blackhole closure

owner :: !b

The owning thread state object

queue :: !b

??

IntClosure
Primitive Int

Fields

ptipe :: PrimType

intVal :: !Int

WordClosure
Primitive Word

Fields

ptipe :: PrimType

wordVal :: !Word

Int64Closure
Primitive Int64

Fields

ptipe :: PrimType

int64Val :: !Int64

Word64Closure
Primitive Word64

Fields

ptipe :: PrimType

word64Val :: !Word64

AddrClosure
Primitive Addr

Fields

ptipe :: PrimType

addrVal :: !Int

FloatClosure
Primitive Float

Fields

ptipe :: PrimType

floatVal :: !Float

DoubleClosure
Primitive Double

Fields

ptipe :: PrimType

doubleVal :: !Double

OtherClosure
Another kind of closure

Fields

info :: !StgInfoTable

hvalues :: ![b]

rawWords :: ![Word]

UnsupportedClosure

Fields

info :: !StgInfoTable

data ClosureType

Constructors

INVALID_OBJECT

CONSTR

CONSTR_1_0

CONSTR_0_1

CONSTR_2_0

CONSTR_1_1

CONSTR_0_2

CONSTR_NOCAF

FUN

FUN_1_0

FUN_0_1

FUN_2_0

FUN_1_1

FUN_0_2

FUN_STATIC

THUNK

THUNK_1_0

THUNK_0_1

THUNK_2_0

THUNK_1_1

THUNK_0_2

THUNK_STATIC

THUNK_SELECTOR

BCO

AP

PAP

AP_STACK

IND

IND_STATIC

RET_BCO

RET_SMALL

RET_BIG

RET_FUN

UPDATE_FRAME

CATCH_FRAME

UNDERFLOW_FRAME

STOP_FRAME

BLOCKING_QUEUE

BLACKHOLE

MVAR_CLEAN

MVAR_DIRTY

TVAR

ARR_WORDS

MUT_ARR_PTRS_CLEAN

MUT_ARR_PTRS_DIRTY

MUT_ARR_PTRS_FROZEN_DIRTY

MUT_ARR_PTRS_FROZEN_CLEAN

MUT_VAR_CLEAN

MUT_VAR_DIRTY

WEAK

PRIM

MUT_PRIM

TSO

STACK

TREC_CHUNK

ATOMICALLY_FRAME

CATCH_RETRY_FRAME

CATCH_STM_FRAME

WHITEHOLE

SMALL_MUT_ARR_PTRS_CLEAN

SMALL_MUT_ARR_PTRS_DIRTY

SMALL_MUT_ARR_PTRS_FROZEN_DIRTY

SMALL_MUT_ARR_PTRS_FROZEN_CLEAN

COMPACT_NFDATA

N_CLOSURE_TYPES

data PrimType

Constructors

PInt

PWord

PInt64

PWord64

PAddr

PFloat

PDouble

class HasHeapRep (a :: TYPE rep) where

Methods

getClosureData :: a -> IO Closure

Info Table types

data StgInfoTable

This is a somewhat faithful representation of an info table. See http://hackage.haskell.org/trac/ghc/browser/includes/rts/storage/InfoTables.h for more details on this data structure.

Constructors

StgInfoTable

Fields

entry :: Maybe EntryFunPtr

ptrs :: HalfWord

nptrs :: HalfWord

tipe :: ClosureType

srtlen :: HalfWord

code :: Maybe ItblCodes

type EntryFunPtr = FunPtr (Ptr () -> IO (Ptr ()))

type HalfWord = Word32

type ItblCodes = Either [Word8] [Word32]

itblSize :: Int

Size in bytes of a standard InfoTable

peekItbl :: Ptr StgInfoTable -> IO StgInfoTable

Read an InfoTable from the heap into a haskell type. WARNING: This code assumes it is passed a pointer to a "standard" info table. If tables_next_to_code is enabled, it will look 1 byte before the start for the entry field.

pokeItbl :: Ptr StgInfoTable -> StgInfoTable -> IO ()

Closure inspection

getBoxedClosureData :: Box -> IO Closure

Like getClosureData, but taking a Box, so it is easier to work with.

allClosures :: GenClosure b -> [b]

For generic code, this function returns all referenced closures.

Boxes

data Box

An arbitrary Haskell value in a safe Box. The point is that even unevaluated thunks can safely be moved around inside the Box, and when required, e.g. in getBoxedClosureData, the function knows how far it has to evaluate the argument.

Constructors

Box Any

asBox :: a -> Box

This takes an arbitrary value and puts it into a box. Note that calls like

asBox (head list)

will put the thunk "head list" into the box, not the element at the head of the list. For that, use careful case expressions:

case list of x:_ -> asBox x

areBoxesEqual :: Box -> Box -> IO Bool

Boxes can be compared, but this is not pure, as different heap objects can, after garbage collection, become the same object.